+++ bin/tests/003_HOSTED_PAGE/1_frontend/0301-HOSTED_REDIRECT_ON_FRONTEND.js
@@ -49,7 +49,7 @@ casper.test.begin('Test Checkout ' + paymentType + ' with redirect', function (t
             checkoutMod.selectMethodInCheckout(test, labelPayByCard, true);
         })
         .then(function () {
-            paymentLibHiPay.fillPaymentFormularByPaymentProduct(currentBrandCC, test);
+            paymentLibHiPay.fillCCFormular( test, false);
         })
         .then(function () {
             adminMod.orderResultSuccess(test);