+++ src/hipay_enterprise/classes/helper/HipayConfigFormHandler.php
@@ -63,9 +63,7 @@ class HipayConfigFormHandler
                         Tools::getValue("api_password_sandbox") &&
                         !Tools::getValue("api_username_sandbox"))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If sandbox api username is filled sandbox api password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If sandbox api username is filled sandbox api password is mandatory");
                     return false;
                 } elseif (($key == "api_tokenjs_username_sandbox" &&
                         Tools::getValue("api_tokenjs_username_sandbox") &&
@@ -74,9 +72,7 @@ class HipayConfigFormHandler
                         Tools::getValue("api_tokenjs_password_publickey_sandbox") &&
                         !Tools::getValue("api_tokenjs_username_sandbox"))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If sandbox api TokenJS username is filled sandbox api TokenJS password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If sandbox api TokenJS username is filled sandbox api TokenJS password is mandatory");
                     return false;
                 } elseif (($key == "api_moto_username_sandbox" &&
                         Tools::getValue("api_moto_username_sandbox") &&
@@ -88,9 +84,7 @@ class HipayConfigFormHandler
                             "api_moto_username_sandbox"
                         ))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If sandbox api MO/TO username is filled sandbox api MO/TO password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If sandbox api MO/TO username is filled sandbox api MO/TO password is mandatory");
                     return false;
                 } else {
                     if ($key == "api_secret_passphrase_sandbox" || $key == "api_moto_secret_passphrase_sandbox") {
@@ -112,9 +106,7 @@ class HipayConfigFormHandler
                         Tools::getValue("api_password_production") &&
                         !Tools::getValue("api_username_production"))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If production api username is filled production api password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If production api username is filled production api password is mandatory");
                     return false;
                 } elseif (($key == "api_tokenjs_username_production" &&
                         Tools::getValue("api_tokenjs_username_production") &&
@@ -123,9 +115,7 @@ class HipayConfigFormHandler
                         Tools::getValue("api_tokenjs_password_publickey_production") &&
                         !Tools::getValue("api_tokenjs_username_production"))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If production api TokenJS username is filled production api TokenJS password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If production api TokenJS username is filled production api TokenJS password is mandatory");
                     return false;
                 } elseif (($key == "api_moto_username_production" &&
                         Tools::getValue("api_moto_username_production") &&
@@ -135,9 +125,7 @@ class HipayConfigFormHandler
                         Tools::getValue("api_moto_password_production") &&
                         !Tools::getValue("api_moto_username_production"))
                 ) {
-                    $this->module->_errors[] = $this->module->l(
-                        "If production api MO/TO username is filled production api MO/TO password is mandatory"
-                    );
+                    $this->module->_errors[] = $this->module->l("If production api MO/TO username is filled production api MO/TO password is mandatory");
                     return false;
                 } else {
                     if ($key == "api_secret_passphrase_production" || $key == "api_moto_secret_passphrase_production") {